+++ src/node/kernel_mempool.cpp
@@ -42,4 +42,12 @@ void KernelMempool::check(const CCoinsViewCache& active_coins_tip, int64_t spend
     m_mempool.check(active_coins_tip, spendheight);
 }
 
+void KernelMempool::MaybeUpdateMempoolForReorg(Chainstate& active_chainstate, DisconnectedBlockTransactions& disconnectpool, bool fAddToMempool)
+{
+    LOCK(::cs_main);
+    LOCK(m_mempool.cs);
+    m_mempool.MaybeUpdateMempoolForReorg(active_chainstate, disconnectpool, fAddToMempool);
+}
+
+
 } // namespace node
